requirements.txt
asgiref==3.8.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
pillow==11.3.0
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
